Add support for InputDeviceInfo https://bugs.webkit.org/show_bug.cgi?id=253896 rdar://problem/106716244
Reviewed by Eric Carlson.

Expose InputDeviceInfo to help web applications sleect the right device and configuration.
To do this, UIProcess is now computing capabilities by creating a source when enumerating each device.
It is sent through IPC to WebProcess that will create the InputDeviceInfo object.
We update IDL to expose InputDeviceInfo.

Covered by WPT tests.

* LayoutTests/fast/mediastream/media-device-info.html:
* LayoutTests/imported/w3c/web-platform-tests/mediacapture-streams/MediaDevices-enumerateDevices.https-expected.txt:
* LayoutTests/imported/w3c/web-platform-tests/mediacapture-streams/MediaStreamTrack-getCapabilities.https-expected.txt:
* LayoutTests/imported/w3c/web-platform-tests/mediacapture-streams/idlharness.https.window-expected.txt:
* LayoutTests/platform/glib/TestExpectations:
* Source/WebCore/CMakeLists.txt:
* Source/WebCore/DerivedSources-input.xcfilelist:
* Source/WebCore/DerivedSources-output.xcfilelist:
* Source/WebCore/DerivedSources.make:
* Source/WebCore/Headers.cmake:
* Source/WebCore/Modules/mediastream/InputDeviceInfo.cpp: Added.
(WebCore::InputDeviceInfo::InputDeviceInfo):
(WebCore::InputDeviceInfo::getCapabilities const):
* Source/WebCore/Modules/mediastream/InputDeviceInfo.h: Added.
* Source/WebCore/Modules/mediastream/InputDeviceInfo.idl: Added.
* Source/WebCore/Modules/mediastream/MediaDeviceInfo.cpp:
(WebCore::toMediaDeviceInfoKind):
* Source/WebCore/Modules/mediastream/MediaDeviceInfo.h:
(): Deleted.
* Source/WebCore/Modules/mediastream/MediaDeviceInfo.idl:
* Source/WebCore/Modules/mediastream/MediaDevices.cpp:
(WebCore::MediaDevices::exposeDevices):
(WebCore::MediaDevices::enumerateDevices):
(WebCore::toMediaDeviceInfoKind): Deleted.
* Source/WebCore/Modules/mediastream/MediaDevices.h:
* Source/WebCore/Modules/mediastream/MediaDevices.idl:
* Source/WebCore/Modules/mediastream/MediaStreamTrack.cpp:
(WebCore::MediaStreamTrack::getCapabilities const):
(WebCore::capabilityDoubleRange): Deleted.
(WebCore::capabilityIntRange): Deleted.
(WebCore::capabilityStringVector): Deleted.
(WebCore::capabilityBooleanVector): Deleted.
* Source/WebCore/Modules/mediastream/MediaStreamTrack.h:
* Source/WebCore/Modules/mediastream/MediaStreamTrack.idl:
* Source/WebCore/Modules/mediastream/MediaTrackCapabilities.cpp: Added.
(WebCore::capabilityDoubleRange):
(WebCore::capabilityIntRange):
(WebCore::capabilityStringVector):
(WebCore::capabilityBooleanVector):
(WebCore::toMediaTrackCapabilities):
* Source/WebCore/Modules/mediastream/MediaTrackCapabilities.h: Added.
* Source/WebCore/Modules/mediastream/MediaTrackCapabilities.idl: Added.
* Source/WebCore/Modules/mediastream/UserMediaClient.h:
* Source/WebCore/Sources.txt:
* Source/WebCore/WebCore.xcodeproj/project.pbxproj:
* Source/WebCore/bindings/js/WebCoreBuiltinNames.h:
* Source/WebCore/platform/mediastream/CaptureDeviceWithCapabilities.h: Added.
* Source/WebCore/platform/mediastream/RealtimeMediaSourceCenter.cpp:
(WebCore::RealtimeMediaSourceCenter::getCapabilities):
* Source/WebCore/platform/mediastream/RealtimeMediaSourceCenter.h:
* Source/WebKit/Shared/WebCoreArgumentCoders.serialization.in:
* Source/WebKit/UIProcess/UserMediaPermissionRequestManagerProxy.cpp:
(WebKit::haveMicrophoneDevice):
(WebKit::UserMediaPermissionRequestManagerProxy::platformGetMediaStreamDevices):
(WebKit::UserMediaPermissionRequestManagerProxy::computeFilteredDeviceList):
(WebKit::UserMediaPermissionRequestManagerProxy::enumerateMediaDevicesForFrame):
* Source/WebKit/UIProcess/UserMediaPermissionRequestManagerProxy.h:
* Source/WebKit/UIProcess/WebPageProxy.cpp:
(WebKit::WebPageProxy::enumerateMediaDevicesForFrame):
* Source/WebKit/UIProcess/WebPageProxy.h:
* Source/WebKit/UIProcess/WebPageProxy.messages.in:
* Source/WebKit/UIProcess/glib/UserMediaPermissionRequestManagerProxyGLib.cpp:
(WebKit::UserMediaPermissionRequestManagerProxy::platformGetMediaStreamDevices):
* Source/WebKit/WebProcess/MediaStream/UserMediaPermissionRequestManager.cpp:
(WebKit::UserMediaPermissionRequestManager::enumerateMediaDevices):
* Source/WebKit/WebProcess/MediaStream/UserMediaPermissionRequestManager.h:

Canonical link: https://commits.webkit.org/261745@main
