+++ LayoutTests/imported/w3c/web-platform-tests/mediacapture-streams/MediaDevices-enumerateDevices.https-expected.txt
@@ -5,5 +5,5 @@ This test checks for the presence of the navigator.mediaDevices.enumerateDevices
 
 PASS mediaDevices.enumerateDevices() is present and working - before capture
 PASS mediaDevices.enumerateDevices() is present and working - after capture
-FAIL InputDeviceInfo is supported promise_test: Unhandled rejection with value: object "ReferenceError: Can't find variable: InputDeviceInfo"
+PASS InputDeviceInfo is supported
