+++ LayoutTests/imported/w3c/web-platform-tests/mediacapture-streams/MediaStreamTrack-getCapabilities.https-expected.txt
@@ -16,23 +16,23 @@ PASS Setup video MediaStreamTrack getCapabilities() test for facingMode
 PASS Setup video MediaStreamTrack getCapabilities() test for resizeMode
 PASS Setup video MediaStreamTrack getCapabilities() test for deviceId
 PASS Setup video MediaStreamTrack getCapabilities() test for groupId
-FAIL Setup audio InputDeviceInfo getCapabilities() test for sampleRate assert_inherits: property "getCapabilities" not found in prototype chain
-FAIL Setup audio InputDeviceInfo getCapabilities() test for sampleSize assert_inherits: property "getCapabilities" not found in prototype chain
-FAIL Setup audio InputDeviceInfo getCapabilities() test for echoCancellation assert_inherits: property "getCapabilities" not found in prototype chain
-FAIL Setup audio InputDeviceInfo getCapabilities() test for autoGainControl assert_inherits: property "getCapabilities" not found in prototype chain
-FAIL Setup audio InputDeviceInfo getCapabilities() test for noiseSuppression assert_inherits: property "getCapabilities" not found in prototype chain
-FAIL Setup audio InputDeviceInfo getCapabilities() test for latency assert_inherits: property "getCapabilities" not found in prototype chain
-FAIL Setup audio InputDeviceInfo getCapabilities() test for channelCount assert_inherits: property "getCapabilities" not found in prototype chain
-FAIL Setup audio InputDeviceInfo getCapabilities() test for deviceId assert_inherits: property "getCapabilities" not found in prototype chain
-FAIL Setup audio InputDeviceInfo getCapabilities() test for groupId assert_inherits: property "getCapabilities" not found in prototype chain
-FAIL Setup video InputDeviceInfo getCapabilities() test for width assert_inherits: property "getCapabilities" not found in prototype chain
-FAIL Setup video InputDeviceInfo getCapabilities() test for height assert_inherits: property "getCapabilities" not found in prototype chain
-FAIL Setup video InputDeviceInfo getCapabilities() test for aspectRatio assert_inherits: property "getCapabilities" not found in prototype chain
-FAIL Setup video InputDeviceInfo getCapabilities() test for frameRate assert_inherits: property "getCapabilities" not found in prototype chain
-FAIL Setup video InputDeviceInfo getCapabilities() test for facingMode assert_inherits: property "getCapabilities" not found in prototype chain
-FAIL Setup video InputDeviceInfo getCapabilities() test for resizeMode assert_inherits: property "getCapabilities" not found in prototype chain
-FAIL Setup video InputDeviceInfo getCapabilities() test for deviceId assert_inherits: property "getCapabilities" not found in prototype chain
-FAIL Setup video InputDeviceInfo getCapabilities() test for groupId assert_inherits: property "getCapabilities" not found in prototype chain
+PASS Setup audio InputDeviceInfo getCapabilities() test for sampleRate
+PASS Setup audio InputDeviceInfo getCapabilities() test for sampleSize
+PASS Setup audio InputDeviceInfo getCapabilities() test for echoCancellation
+PASS Setup audio InputDeviceInfo getCapabilities() test for autoGainControl
+PASS Setup audio InputDeviceInfo getCapabilities() test for noiseSuppression
+PASS Setup audio InputDeviceInfo getCapabilities() test for latency
+PASS Setup audio InputDeviceInfo getCapabilities() test for channelCount
+PASS Setup audio InputDeviceInfo getCapabilities() test for deviceId
+PASS Setup audio InputDeviceInfo getCapabilities() test for groupId
+PASS Setup video InputDeviceInfo getCapabilities() test for width
+PASS Setup video InputDeviceInfo getCapabilities() test for height
+PASS Setup video InputDeviceInfo getCapabilities() test for aspectRatio
+PASS Setup video InputDeviceInfo getCapabilities() test for frameRate
+PASS Setup video InputDeviceInfo getCapabilities() test for facingMode
+PASS Setup video InputDeviceInfo getCapabilities() test for resizeMode
+PASS Setup video InputDeviceInfo getCapabilities() test for deviceId
+PASS Setup video InputDeviceInfo getCapabilities() test for groupId
 PASS Audio track getCapabilities() sampleRate property present.
 PASS Audio track getCapabilities() sampleRate properly supported.
 FAIL Audio track getCapabilities() sampleSize property present. assert_true: expected true got false
@@ -69,4 +69,40 @@ PASS Video track getCapabilities() deviceId property present.
 PASS Video track getCapabilities() deviceId properly supported.
 FAIL Video track getCapabilities() groupId property present. assert_true: expected true got false
 FAIL Video track getCapabilities() groupId properly supported. assert_equals: expected "string" but got "undefined"
+PASS Audio device getCapabilities() sampleRate property present.
+PASS Audio device getCapabilities() sampleRate properly supported.
+FAIL Audio device getCapabilities() sampleSize property present. assert_true: expected true got false
+FAIL Audio device getCapabilities() sampleSize properly supported. assert_equals: expected "object" but got "undefined"
+PASS Audio device getCapabilities() echoCancellation property present.
+PASS Audio device getCapabilities() echoCancellation properly supported.
+FAIL Audio device getCapabilities() autoGainControl property present. assert_true: expected true got false
+FAIL Audio device getCapabilities() autoGainControl properly supported. undefined is not an object (evaluating 'capability.length')
+FAIL Audio device getCapabilities() noiseSuppression property present. assert_true: expected true got false
+FAIL Audio device getCapabilities() noiseSuppression properly supported. undefined is not an object (evaluating 'capability.length')
+FAIL Audio device getCapabilities() latency property present. assert_true: expected true got false
+FAIL Audio device getCapabilities() latency properly supported. assert_equals: expected "object" but got "undefined"
+FAIL Audio device getCapabilities() channelCount property present. assert_true: expected true got false
+FAIL Audio device getCapabilities() channelCount properly supported. assert_equals: expected "object" but got "undefined"
+PASS Audio device getCapabilities() deviceId property present.
+PASS Audio device getCapabilities() deviceId properly supported.
+FAIL Audio device getCapabilities() groupId property present. assert_true: expected true got false
+FAIL Audio device getCapabilities() groupId properly supported. assert_equals: expected "string" but got "undefined"
+PASS Video device getCapabilities() width property present.
+PASS Video device getCapabilities() width properly supported.
+PASS Video device getCapabilities() height property present.
+PASS Video device getCapabilities() height properly supported.
+PASS Video device getCapabilities() aspectRatio property present.
+PASS Video device getCapabilities() aspectRatio properly supported.
+PASS Video device getCapabilities() frameRate property present.
+PASS Video device getCapabilities() frameRate properly supported.
+PASS Video device getCapabilities() facingMode property present.
+PASS Video device getCapabilities() facingMode properly supported.
+FAIL Video device getCapabilities() resizeMode property present. assert_true: expected true got false
+FAIL Video device getCapabilities() resizeMode properly supported. undefined is not an object (evaluating 'capability.forEach')
+FAIL Video device getCapabilities() resizeMode properly supported. Value: none undefined is not an object (evaluating 'expected.indexOf')
+FAIL Video device getCapabilities() resizeMode properly supported. Value: crop-and-scale undefined is not an object (evaluating 'expected.indexOf')
+PASS Video device getCapabilities() deviceId property present.
+PASS Video device getCapabilities() deviceId properly supported.
+FAIL Video device getCapabilities() groupId property present. assert_true: expected true got false
+FAIL Video device getCapabilities() groupId properly supported. assert_equals: expected "string" but got "undefined"
