+++ Source/WebCore/Modules/mediastream/MediaDeviceInfo.cpp
@@ -28,8 +28,6 @@
 
 #if ENABLE(MEDIA_STREAM)
 
-#include <wtf/IsoMallocInlines.h>
-
 namespace WebCore {
 
 WTF_MAKE_ISO_ALLOCATED_IMPL(MediaDeviceInfo);
@@ -45,6 +43,25 @@ MediaDeviceInfo::MediaDeviceInfo(const String& label, const String& deviceId, co
     , m_groupId(groupId)
     , m_kind(kind)
 {
+}
+
+MediaDeviceInfo::Kind toMediaDeviceInfoKind(CaptureDevice::DeviceType type)
+{
+    switch (type) {
+    case CaptureDevice::DeviceType::Microphone:
+        return MediaDeviceInfo::Kind::Audioinput;
+    case CaptureDevice::DeviceType::Speaker:
+        return MediaDeviceInfo::Kind::Audiooutput;
+    case CaptureDevice::DeviceType::Camera:
+    case CaptureDevice::DeviceType::Screen:
+    case CaptureDevice::DeviceType::Window:
+        return MediaDeviceInfo::Kind::Videoinput;
+    case CaptureDevice::DeviceType::SystemAudio:
+    case CaptureDevice::DeviceType::Unknown:
+        break;
+    }
+    ASSERT_NOT_REACHED();
+    return MediaDeviceInfo::Kind::Audioinput;
 }
 
 } // namespace WebCore
