+++ Source/WebCore/Modules/mediastream/MediaDevices.h
@@ -48,6 +48,7 @@
 namespace WebCore {
 
 class Document;
+class InputDeviceInfo;
 class MediaDeviceInfo;
 class MediaStream;
 class UserGestureToken;
@@ -66,7 +67,7 @@ public:
     Document* document() const;
 
     using Promise = DOMPromiseDeferred<IDLInterface<MediaStream>>;
-    using EnumerateDevicesPromise = DOMPromiseDeferred<IDLSequence<IDLInterface<MediaDeviceInfo>>>;
+    using EnumerateDevicesPromise = DOMPromiseDeferred<IDLSequence<IDLUnion<IDLInterface<MediaDeviceInfo>, IDLInterface<InputDeviceInfo>>>>;
 
     enum class DisplayCaptureSurfaceType {
         Monitor,
@@ -101,7 +102,7 @@ private:
     void scheduledEventTimerFired();
     bool addEventListener(const AtomString& eventType, Ref<EventListener>&&, const AddEventListenerOptions&) override;
 
-    void exposeDevices(const Vector<CaptureDevice>&, MediaDeviceHashSalts&&, EnumerateDevicesPromise&&);
+    void exposeDevices(Vector<CaptureDeviceWithCapabilities>&&, MediaDeviceHashSalts&&, EnumerateDevicesPromise&&);
     void listenForDeviceChanges();
 
     friend class JSMediaDevicesOwner;
