+++ Source/WebCore/Modules/mediastream/MediaDevices.idl
@@ -28,6 +28,8 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
+typedef (MediaDeviceInfo or InputDeviceInfo) DeviceInfo;
+
 [
     ActiveDOMObject,
     Conditional=MEDIA_STREAM,
@@ -35,7 +37,7 @@
     Exposed=Window
 ] interface MediaDevices : EventTarget {
     attribute EventHandler ondevicechange;
-    Promise<sequence<MediaDeviceInfo>> enumerateDevices();
+    Promise<sequence<DeviceInfo>> enumerateDevices();
 
     MediaTrackSupportedConstraints getSupportedConstraints();
     [PrivateIdentifier, PublicIdentifier] Promise<MediaStream> getUserMedia(optional MediaStreamConstraints constraints);
