+++ Source/WebCore/Modules/mediastream/MediaTrackCapabilities.h
@@ -1,0 +1,58 @@
+/*
+ * Copyright (C) 2023 Apple Inc. All rights reserved.
+ *
+ * Redistribution and use in source and binary forms, with or without
+ * modification, are permitted provided that the following conditions
+ * are met:
+ * 1.  Redistributions of source code must retain the above copyright
+ *     notice, this list of conditions and the following disclaimer.
+ * 2.  Redistributions in binary form must reproduce the above copyright
+ *     notice, this list of conditions and the following disclaimer in the
+ *     documentation and/or other materials provided with the distribution.
+ *
+ * THIS SOFTWARE IS PROVIDED BY APPLE INC. AND ITS CONTRIBUTORS ``AS IS'' AND ANY
+ * EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED
+ * WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
+ * DISCLAIMED. IN NO EVENT SHALL APPLE INC. OR ITS CONTRIBUTORS BE LIABLE FOR ANY
+ * DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES
+ * (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES;
+ * LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON
+ * ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
+ * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
+ * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
+ */
+
+#pragma once
+
+#if ENABLE(MEDIA_STREAM)
+
+#include "DoubleRange.h"
+#include "LongRange.h"
+#include <wtf/Vector.h>
+#include <wtf/text/WTFString.h>
+
+namespace WebCore {
+
+class RealtimeMediaSourceCapabilities;
+
+struct MediaTrackCapabilities {
+    std::optional<LongRange> width;
+    std::optional<LongRange> height;
+    std::optional<DoubleRange> aspectRatio;
+    std::optional<DoubleRange> frameRate;
+    std::optional<Vector<String>> facingMode;
+    std::optional<DoubleRange> volume;
+    std::optional<LongRange> sampleRate;
+    std::optional<LongRange> sampleSize;
+    std::optional<Vector<bool>> echoCancellation;
+    String deviceId;
+    String groupId;
+    String displaySurface;
+    std::optional<DoubleRange> focusDistance;
+    std::optional<DoubleRange> zoom;
+};
+
+MediaTrackCapabilities toMediaTrackCapabilities(const RealtimeMediaSourceCapabilities&);
+} // namespace WebCore
+
+#endif // ENABLE(MEDIA_STREAM)
