+++ Source/WebKit/UIProcess/UserMediaPermissionRequestManagerProxy.cpp
@@ -34,6 +34,7 @@
 #include "WebProcessPool.h"
 #include "WebProcessProxy.h"
 #include "WebsiteDataStore.h"
+#include <WebCore/CaptureDeviceWithCapabilities.h>
 #include <WebCore/MediaConstraints.h>
 #include <WebCore/MockRealtimeMediaSourceCenter.h>
 #include <WebCore/PlatformMediaSessionManager.h>
@@ -786,26 +787,35 @@ bool UserMediaPermissionRequestManagerProxy::wasGrantedVideoOrAudioAccess(FrameI
     return m_grantedFrames.contains(frameID);
 }
 
-static inline bool haveMicrophoneDevice(const Vector<WebCore::CaptureDevice>& devices, const String& deviceID)
+static inline bool haveMicrophoneDevice(const Vector<CaptureDeviceWithCapabilities>& devices, const String& deviceID)
 {
-    return std::any_of(devices.begin(), devices.end(), [&deviceID](auto& device) {
+    return std::any_of(devices.begin(), devices.end(), [&deviceID](auto& deviceWithCapabilities) {
+        auto& device = deviceWithCapabilities.device;
         return device.persistentId() == deviceID && device.type() == CaptureDevice::DeviceType::Microphone;
     });
 }
 
 #if !USE(GLIB)
-void UserMediaPermissionRequestManagerProxy::platformGetMediaStreamDevices(CompletionHandler<void(Vector<WebCore::CaptureDevice>&&)>&& completionHandler)
+void UserMediaPermissionRequestManagerProxy::platformGetMediaStreamDevices(bool revealIdsAndLabels, CompletionHandler<void(Vector<CaptureDeviceWithCapabilities>&&)>&& completionHandler)
 {
-    RealtimeMediaSourceCenter::singleton().getMediaStreamDevices(WTFMove(completionHandler));
+    RealtimeMediaSourceCenter::singleton().getMediaStreamDevices([revealIdsAndLabels, completionHandler = WTFMove(completionHandler)](auto&& devices) mutable {
+        auto deviceWithCapabilities = map(devices, [revealIdsAndLabels](auto&& device) -> CaptureDeviceWithCapabilities {
+            RealtimeMediaSourceCapabilities capabilities;
+            if (revealIdsAndLabels)
+                capabilities = RealtimeMediaSourceCenter::singleton().getCapabilities(device);
+            return { WTFMove(device), WTFMove(capabilities) };
+        });
+        completionHandler(WTFMove(deviceWithCapabilities));
+    });
 }
 #endif
 
-void UserMediaPermissionRequestManagerProxy::computeFilteredDeviceList(bool revealIdsAndLabels, CompletionHandler<void(Vector<CaptureDevice>&&)>&& completion)
+void UserMediaPermissionRequestManagerProxy::computeFilteredDeviceList(bool revealIdsAndLabels, CompletionHandler<void(Vector<CaptureDeviceWithCapabilities>&&)>&& completion)
 {
     static const unsigned defaultMaximumCameraCount = 1;
     static const unsigned defaultMaximumMicrophoneCount = 1;
 
-    platformGetMediaStreamDevices([logIdentifier = LOGIDENTIFIER, this, weakThis = WeakPtr { *this }, revealIdsAndLabels, completion = WTFMove(completion)](auto&& devices) mutable {
+    platformGetMediaStreamDevices(revealIdsAndLabels, [logIdentifier = LOGIDENTIFIER, this, weakThis = WeakPtr { *this }, revealIdsAndLabels, completion = WTFMove(completion)](auto&& devicesWithCapabilities) mutable {
 
         if (!weakThis) {
             completion({ });
@@ -818,8 +828,9 @@ void UserMediaPermissionRequestManagerProxy::computeFilteredDeviceList(bool reve
         bool hasCamera = false;
         bool hasMicrophone = false;
 
-        Vector<CaptureDevice> filteredDevices;
-        for (const auto& device : devices) {
+        Vector<CaptureDeviceWithCapabilities> filteredDevices;
+        for (auto& deviceWithCapabilities : devicesWithCapabilities) {
+            auto& device = deviceWithCapabilities.device;
             if (!device.enabled() || (device.type() != WebCore::CaptureDevice::DeviceType::Camera && device.type() != WebCore::CaptureDevice::DeviceType::Microphone && device.type() != WebCore::CaptureDevice::DeviceType::Speaker))
                 continue;
             hasCamera |= device.type() == WebCore::CaptureDevice::DeviceType::Camera;
@@ -833,11 +844,14 @@ void UserMediaPermissionRequestManagerProxy::computeFilteredDeviceList(bool reve
                     continue;
             } else {
                 // We only expose speakers tied to a microphone for the moment.
-                if (device.type() == WebCore::CaptureDevice::DeviceType::Speaker && !haveMicrophoneDevice(devices, device.groupId()))
+                if (device.type() == WebCore::CaptureDevice::DeviceType::Speaker && !haveMicrophoneDevice(devicesWithCapabilities, device.groupId()))
                     continue;
             }
 
-            filteredDevices.append(revealIdsAndLabels ? device : CaptureDevice({ }, device.type(), { }, { }));
+            if (revealIdsAndLabels)
+                filteredDevices.append(deviceWithCapabilities);
+            else
+                filteredDevices.append({ { { }, device.type(), { }, { } }, { } });
         }
 
         m_hasFilteredDeviceList = !revealIdsAndLabels;
@@ -848,7 +862,7 @@ void UserMediaPermissionRequestManagerProxy::computeFilteredDeviceList(bool reve
 }
 #endif
 
-void UserMediaPermissionRequestManagerProxy::enumerateMediaDevicesForFrame(FrameIdentifier frameID, Ref<SecurityOrigin>&& userMediaDocumentOrigin, Ref<SecurityOrigin>&& topLevelDocumentOrigin, CompletionHandler<void(const Vector<CaptureDevice>&, MediaDeviceHashSalts&&)>&& completionHandler)
+void UserMediaPermissionRequestManagerProxy::enumerateMediaDevicesForFrame(FrameIdentifier frameID, Ref<SecurityOrigin>&& userMediaDocumentOrigin, Ref<SecurityOrigin>&& topLevelDocumentOrigin, CompletionHandler<void(const Vector<CaptureDeviceWithCapabilities>&, MediaDeviceHashSalts&&)>&& completionHandler)
 {
 #if ENABLE(MEDIA_STREAM)
     ALWAYS_LOG(LOGIDENTIFIER);
